+++ braintree/test/nonces.py
@@ -9,6 +9,7 @@ class Nonces(object):
     ApplePayAmEx = "fake-apple-pay-amex-nonce"
     AbstractTransactable = "fake-abstract-transactable-nonce"
     Europe = "fake-europe-bank-account-nonce"
+    # NEXT_MAJOR_VERSION - rename AndroidPay to GooglePay
     AndroidPayCard = "fake-android-pay-nonce"
     AndroidPayCardDiscover = "fake-android-pay-discover-nonce"
     AndroidPayCardVisa = "fake-android-pay-visa-nonce"