+++ braintree/util/constants.py
@@ -1,3 +1,4 @@
+#NEXT_MAJOR_VERSION we can change all Constants to enums, not sure if we'll still need this
 class Constants(object):
     @staticmethod
     def get_all_constant_values_from_class(klass):